Task events in Topcased

+++ nakeduml-runtime/nakeduml-runtime-domain/src/main/java/org/nakeduml/runtime/domain/IProcessStep.java
@@ -2,7 +2,7 @@ package org.nakeduml.runtime.domain;
 
 
 public interface IProcessStep {
-	public int getId();
+	public long getId();
 	public String getQualifiedName();
 	public String getHumanName();
 	public TriggerMethod[] getTriggerMethods();
